add admin page for admidat actions
