+++ app/controllers/offers_controller.rb
@@ -9,9 +9,6 @@ class OffersController < ApplicationController
       format.html
       format.json { render json: @offers.map(&:name) }
     end
-    autorisation_admitad
-    get_action_data
-    rec_user_actions if cookies[:action_data].present?
   end
 
   def offer_redirect
