+++ app/controllers/users_controller.rb
@@ -5,7 +5,7 @@ class UsersController < ApplicationController
   before_action :authenticate_user!
   before_action :find_transactions, only: %i[show filter_status]
 
-  def show;  end
+  def show; end
 
   def create_withdrawal_request
     Transaction.where(user_id: current_user.id, status: 0).each do |transaction|
@@ -31,10 +31,10 @@ class UsersController < ApplicationController
         withdrawal_id = "1000#{Time.now.strftime('%I%M%S')}".to_i
         amount = current_user.transactions.where(status: 1).sum(:cashback_sum)
         Transaction.create(user_id: current_user.id,
-                               action_id: withdrawal_id,
-                               cashback_sum: amount,
-                               status: 5,
-                               offer_id: Offer.first.id).save
+                           action_id: withdrawal_id,
+                           cashback_sum: amount,
+                           status: 5,
+                           offer_id: Offer.first.id).save
 
         liqpay.api('request', {
                      action: 'p2pcredit',
@@ -65,18 +65,12 @@ class UsersController < ApplicationController
     error_notice.push('Введіть карту') if params[:card_num].blank?
     error_notice.push('Введіть прізвище') if params[:last_name].blank?
     error_notice.push("Введіть ім\\'я") if params[:first_name].blank?
-    error_notice.push('В прізвищі не можуть бути спец символи та цифри') unless params[:last_name].match(valid_name_regex)
+    unless params[:last_name].match(valid_name_regex)
+      error_notice.push('В прізвищі не можуть бути спец символи та цифри')
+    end
     error_notice.push('В імені не можуть бути спец символи та цифри') unless params[:first_name].match(valid_name_regex)
     error_notice.push('Неправильний номер картки') unless params[:card_num].to_s.scan(/\D/).empty?
     error_notice.push('Закороткий номер картки') unless params[:card_num].length <= 16
     error_notice
   end
-
-  # def sort_column
-  #  Transaction.column_names.include?(params[:sort]) ? params[:sort] : nil
-  # end
-
-  # def sort_direction
-  #  %w[asc desc].include?(params[:direction]) ? params[:direction] : nil
-  # end
 end
