+++ app/views/admins/admin_rec_actions_result.js.erb
@@ -1,0 +1,1 @@
+console.log('click');
