+++ app/views/offers/index.html.erb
@@ -146,14 +146,3 @@
     </div>
   </div>
 <% end %>
-<%#= link_to 'GET CODE', 'https://www.admitad.com/api/authorize/?scope=statistics advcampaigns banners websites&state=7c232ff20e64432fbe071228c0779f&redirect_uri=https://cback.club/&response_type=code&client_id=9Oo9LsDIaQhqCUtVkbSFIPfSmXQ7mQ'%>
-
-<%#= link_to 'GET CODE', 'https://www.admitad.com/api/authorize/?scope=statistics advcampaigns banners websites&state=7c232ff20e64432fbe071228c0779f&redirect_uri=http://127.0.0.1:3000/&response_type=code&client_id=9Oo9LsDIaQhqCUtVkbSFIPfSmXQ7mQ'%>
-<!--<h1>DEBUG</h1>-->
-<!--<div>-->
-<%#= @response %>
-<!--  <h1>________</h1>-->
-<%#=  @subid_data %>
-<!--  <h1>________</h1>-->
-<%#= @action_data %>
-<!--</div>-->
