+++ app/views/users/_withdrawal_liqpay.html.erb
@@ -1,5 +1,4 @@
 <%= form_with url: withdrawal_liqpay_path, id: "withdrawal_liqpay_form", local: false do |form| %>
-<%# if post.errors.any? %>
   <div class="card-wrap">
     <div class="field">
       <%= form.label :card_num %>
