+++ config/routes.rb
@@ -7,18 +7,25 @@ Rails.application.routes.draw do
   get 'offers/index'
   get 'offers/show'
   get 'autocomplete', to: 'offers#autocomplete', as: 'autocomplete'
+  # offer invisible redirect
   get 'offers/offer_redirect', to: 'offers#offer_redirect', as: 'redirect'
-  get 'withdrawal', to:'users#create_withdrawal_request'
+  # filter for users transactions
   get 'filter_status', to: 'users#filter_status', as: 'filter_status'
+  # liqpay routes
+  get 'withdrawal', to: 'users#create_withdrawal_request'
   post 'withdrawal_liqpay', to: 'users#withdrawal_liqpay', as: 'withdrawal_liqpay'
   post 'withdrawal_get', to: 'admins#withdrawal_get', as: 'withdrawal_get'
-
+  # admitad routes
+  get 'admin_panel', to: 'admins#index', as: 'admin_panel'
+  get 'autorisation_admitad', to: 'admins#autorisation_admitad', as: 'autorisation_admitad'
+  post 'get_action_data', to: 'admins#get_action_data', as: 'get_action_data'
 
   devise_for :users,
              controllers: { omniauth_callbacks: 'users/omniauth_callbacks', registrations: 'users/registrations',
                             sessions: 'users/sessions' }
   root 'offers#index'
   get '/agreement', to: 'pages#agreement', as: "agreement"
+  # page error messages
   match "/404", to: "errors#not_found", via: :all
   match "/422", to: "errors#unacceptable", via: :all
   match "/500", to: "errors#internal_server_error", via: :all
